+++ src/main/java/ru/beeline/fdmauth/repository/UserProfileRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 
 
 @Repository
-public interface UserProfileRepository extends JpaRepository<UserProfile, Long> {
+public interface UserProfileRepository extends JpaRepository<UserProfile, Integer> {
     @Query("SELECT DISTINCT up FROM UserProfile up " +
             "JOIN up.userRoles ur " +
             "JOIN ur.role r " +
