+++ src/main/java/ru/beeline/fdmauth/service/RoleService.java
@@ -211,7 +211,7 @@ public class RoleService {
         }
     }
 
-    public void deleteAllByUserProfileId(Long id) {
+    public void deleteAllByUserProfileId(Integer id) {
         userRolesRepository.deleteAllByUserProfileId(id);
     }
